Refactor benchmarking module code
