+++ src/ontime/module/benchmarking/__init__.py
@@ -1,0 +1,3 @@
+from .benchmark_metric import BenchmarkMetric
+from .benchmark_model import AbstractBenchmarkModel, BenchmarkMode
+from .benchmark_dataset import BenchmarkDataset
