+++ src/ontime/module/benchmarking/benchmark_metric.py
@@ -1,0 +1,13 @@
+from ontime.core.time_series.time_series import TimeSeries
+
+class BenchmarkMetric:
+    def __init__(self, name: str, metric_function, reduction = None):
+        self.metric = metric_function
+        self.name = name
+        self.reduction = reduction
+
+    def compute(self, target: TimeSeries, pred: TimeSeries):
+        """
+        Compute the metric on the target and predicted time series.
+        """
+        return self.metric(target, pred, component_reduction=self.reduction)
